+++ desktop/internal/bridge/process_windows.go
@@ -1,0 +1,9 @@
+//go:build windows
+
+package bridge
+
+import "syscall"
+
+func hiddenProcessAttributes() *syscall.SysProcAttr {
+	return &syscall.SysProcAttr{HideWindow: true, CreationFlags: 0x08000000}
+}