+++ desktop/main.go
@@ -1,0 +1,35 @@
+package main
+
+import (
+	"embed"
+
+	"github.com/wailsapp/wails/v2"
+	"github.com/wailsapp/wails/v2/pkg/options"
+	"github.com/wailsapp/wails/v2/pkg/options/assetserver"
+	"github.com/wailsapp/wails/v2/pkg/options/windows"
+)
+
+//go:embed all:frontend/dist
+var assets embed.FS
+
+func main() {
+	app := NewApp()
+	err := wails.Run(&options.App{
+		Title:            "RemoteBridge",
+		Width:            1080,
+		Height:           720,
+		MinWidth:         900,
+		MinHeight:        620,
+		BackgroundColour: &options.RGBA{R: 245, G: 247, B: 251, A: 1},
+		AssetServer:      &assetserver.Options{Assets: assets},
+		OnStartup:        app.startup,
+		Bind:             []interface{}{app},
+		Windows: &windows.Options{
+			Theme:                windows.SystemDefault,
+			WebviewIsTransparent: false,
+		},
+	})
+	if err != nil {
+		println("RemoteBridge failed:", err.Error())
+	}
+}